+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJSJH62HQQ818V6MNVB27HF1"
+        "01KJVN9RY2X3CXFPJ4GKTTQ5DB"
     )
     r"""The unique identifier of the data source"""