+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-03T10:04:52.010Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-04T05:31:45.405Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")