+++ packages/orq-rc/RELEASES.md
@@ -6279,3 +6279,13 @@ Based on:
 - [python v4.5.0-rc.23] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.23] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.23 - packages/orq-rc
+
+## 2026-03-04 05:34:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.740.0 (2.846.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.24] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.24] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.24 - packages/orq-rc